ref: SQLite graph saving now in Graph class
